-Buscar exhaustivamente un ED -Cambio a longs de varias clases

+++ src/Valuador.java
@@ -1,6 +1,6 @@
 public class Valuador {
 
-    public static double CalcularPesoTransicion(double []Ct, int []St){
+    public static double CalcularPesoTransicion(double []Ct, long []St){
         int aux=0;
         for(int i=0; i<St.length; i++){
                 aux += Ct[i] * St[i];
@@ -8,14 +8,14 @@ public class Valuador {
         }
         return aux;
     }
-    public static double CalcularPesoLugar(double[] Cl, int []Sl){
+    public static double CalcularPesoLugar(double[] Cl, long []Sl){
         int aux=0;
         for(int i=0; i<Sl.length; i++){
                 aux += Cl[i] * Sl[i];
         }
         return aux;
     }
-    public static double CalcularPesoTotal(double[] Cl, double []Ct, int[] Sl, int[] St){
+    public static double CalcularPesoTotal(double[] Cl, double []Ct, long[] Sl, long[] St){
         double total=0.0;
         total = CalcularPesoLugar(Cl,Sl) + CalcularPesoTransicion(Ct,St);
         return total;
